requirements.txt
# python==3.6.5
# absl-py==1.3.0
astor==0.8.1
async-timeout==4.0.2
atari-py==0.2.9
attrs==22.1.0
cached-property==1.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
cycler==0.10.0
Cython==0.29.32
dataclasses==0.8
decorator==4.3.0
Deprecated==1.2.13
fasteners==0.15
filelock==3.4.1
flatbuffers==22.9.24
funcsigs==1.0.2
gast==0.5.3
glfw==2.5.5
grpcio==1.48.2
gtimer==1.0.0b5
gym==0.12.0
h5py==3.1.0
idna==3.4
imageio==2.15.0
importlib-metadata==4.8.3
iniconfig==1.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.0.1
lockfile==0.12.2
lz4==3.1.10
Markdown==3.3.7
matplotlib==3.0.2
mock==4.0.3
monotonic==1.5
mujoco-py==1.50.1.68
networkx==2.2
numpy==1.16.0
opencv-python-headless==4.6.0.66
packaging==21.3
Pillow==8.4.0
pluggy==1.0.0
protobuf==3.19.6
psutil==5.9.3
# py==1.11.0
# py-spy==0.3.14
pycparser==2.21
pyglet==1.5.27
PyOpenGL==3.1.6
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.7.5
PyWavelets==1.0.1
PyYAML==6.0
ray==0.6.4
redis==4.3.4
requests==2.27.1
scikit-image==0.14.2
scipy==1.5.4
setproctitle==1.2.3
six==1.16.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
# tensorflow-gpu==1.14.0
# tensorflow-probability==0.7.0
# tensorflow=1.13.1
tensorflow-probability==0.6.0
termcolor==1.1.0
tomli==1.2.3
toolz==0.9.0
tqdm==4.32.1
typing==3.7.4.3
typing_extensions==4.1.1
urllib3==1.26.12
Werkzeug==2.0.3
wrapt==1.14.1
zipp==3.6.0
Serializable@git+https://github.com/hartikainen/serializable.git
# cuda10.0
# mujoco150